requirement.txt
ipdb
joblib
cytoolz
cython
lz4==2.1.9
lmdb==0.97
msgpack-numpy
msgpack
toolz
simplejson
psutil
tensorboard
tqdm
easydict
pycocotools>=2.0.1
opencv-python
tensorboardX==2.0
torchvision